add concrete types to identity fields

+++ messages/auth.messages.go
@@ -6,6 +6,6 @@ import (
 
 type Claims struct {
 	jwt.RegisteredClaims
-	UserId 		uint 	`json:"user_id"`
+	UserId 		uint64 	`json:"user_id"`
 	Username 	string 	`json:"username"`
 }
