+++ repositories/message.sqlite.repository.go
@@ -13,7 +13,7 @@ func NewMessageRepository(db *gorm.DB) *SqliteMessageRepository {
 	return &SqliteMessageRepository{db: db}
 }
 
-func (r *SqliteMessageRepository) SendMessage(text string, fromId, toId uint) error {
+func (r *SqliteMessageRepository) SendMessage(text string, fromId, toId uint64) error {
 	message := &models.Message{
 		Text:   text,
 		FromId: fromId,
@@ -22,7 +22,7 @@ func (r *SqliteMessageRepository) SendMessage(text string, fromId, toId uint) er
 	return r.db.Create(message).Error
 }
 
-func (r *SqliteMessageRepository) GetMessages(fromId, toId uint) ([]*models.Message, error) {
+func (r *SqliteMessageRepository) GetMessages(fromId, toId uint64) ([]*models.Message, error) {
 	var messages []*models.Message
 	if err := r.db.Where("from_id = ? AND to_id = ?", fromId, toId).Find(&messages).Error; err != nil {
 		return nil, err
